Added IRequest interfaces to make it easier to raise requests in Mediator package
Core.Mediator (2.0.0):
-  Changed IMessageHandler.Priority type from ushort? to byte?
- Updated all components to use byte priority
- Added IRequest and IRequest<T> interfaces that requests need to implement before they can be raised

+++ Sels.Core.Mediator/Components/Event/EventSubscriptionManager.cs
@@ -61,7 +61,7 @@ namespace Sels.Core.Mediator.Event
             }
         }
         /// <inheritdoc/>
-        public EventSubscription Subscribe<TEvent>(Delegates.Async.AsyncAction<IEventListenerContext, TEvent, CancellationToken> subscriberAction, ushort? priority = null)
+        public EventSubscription Subscribe<TEvent>(Delegates.Async.AsyncAction<IEventListenerContext, TEvent, CancellationToken> subscriberAction, byte? priority = null)
         {
             using var methodLogger = _logger.TraceMethod(this);
             subscriberAction.ValidateArgument(nameof(subscriberAction));
@@ -84,7 +84,7 @@ namespace Sels.Core.Mediator.Event
             return new EventSubscription(listener, null, () => Unsubscribe(listener));
         }
         /// <inheritdoc/>
-        public EventSubscription Subscribe(Delegates.Async.AsyncAction<IEventListenerContext, object, CancellationToken> subscriberAction, ushort? priority = null)
+        public EventSubscription Subscribe(Delegates.Async.AsyncAction<IEventListenerContext, object, CancellationToken> subscriberAction, byte? priority = null)
         {
             using var methodLogger = _logger.TraceMethod(this);
             subscriberAction.ValidateArgument(nameof(subscriberAction));
@@ -136,7 +136,7 @@ namespace Sels.Core.Mediator.Event
             return new EventSubscription(listener, typeof(TEvent), () => Unsubscribe(listener));
         }
         /// <inheritdoc/>
-        public EventSubscription Subscribe(Delegates.Async.AsyncAction<IEventListenerContext, TEvent, CancellationToken> subscriberAction, ushort? priority = null)
+        public EventSubscription Subscribe(Delegates.Async.AsyncAction<IEventListenerContext, TEvent, CancellationToken> subscriberAction, byte? priority = null)
         {
             using var methodLogger = _logger.TraceMethod(this);
             return Subscribe(new DelegateEventListener<TEvent>(subscriberAction) { Priority = priority });
